src/modules/client-adm/usecase/find-client/find-client.usecase.ts
import ClientGateway from "../../gateway/client.gateway";
import {
  FindClientInputDto,
  FindClientOutputDto,
} from "./find-client.usecase.dto";

export default class FindClientUseCase {
  private _clientRepository: ClientGateway;

  constructor(clientRepository: ClientGateway) {
    this._clientRepository = clientRepository;
  }

  async execute(input: FindClientInputDto): Promise<FindClientOutputDto> {
    const client = await this._clientRepository.find(input.id);

    return {
      id: client.id.id,
      name: client.name,
      email: client.email,
      document: client.document,
      street: client.address.value.street,
      number: client.address.value.number,
      complement: client.address.value.complement,
      city: client.address.value.city,
      state: client.address.value.state,
      zipCode: client.address.value.zipCode,
    };
  }
}
